reorganize airdrop components and add structure for permission modal
