+++ src/App.tsx
@@ -15,8 +15,8 @@ import { useEnvironment } from './hooks/useEnvironment.ts';
 import { SplashScreen } from './containers/SplashScreen';
 import { useMiningEffects } from './hooks/mining/useMiningEffects.ts';
 import { setupLogger } from './utils/logger.ts';
-import AirdropLogin from './containers/Airdrop/AirdropLogin.tsx';
 import { useAirdropSyncState } from './hooks/airdrop/useAirdropSyncState.ts';
+import AirdropLogin from './containers/Airdrop/AirdropLogin/AirdropLogin.tsx';
 
 function App() {
     useAirdropSyncState();
