+++ src/containers/Airdrop/AirdropPermission/AirdropPermission.tsx
@@ -1,0 +1,5 @@
+import { Wrapper } from './styles';
+
+export default function AirdropPermission() {
+    return <Wrapper>Permission</Wrapper>;
+}
