+++ src/containers/Dashboard/SetupView/SetupViewContainer.tsx
@@ -2,6 +2,8 @@ import { useState } from 'react';
 import useAppStateStore from '@app/store/appStateStore';
 import SetupView from './SetupView';
 import { useInterval } from '@app/hooks/useInterval.ts';
+import AirdropPermission from '@app/containers/Airdrop/AirdropPermission/AirdropPermission';
+import { Wrapper } from './styles';
 
 function SetupViewContainer() {
     const setupTitle = useAppStateStore((s) => s.setupTitle);
@@ -15,7 +17,12 @@ function SetupViewContainer() {
 
     const progressPercentage = Math.floor(progress * 100);
 
-    return <SetupView title={setupTitle} progressPercentage={progressPercentage} />;
+    return (
+        <Wrapper>
+            <SetupView title={setupTitle} progressPercentage={progressPercentage} />
+            <AirdropPermission />
+        </Wrapper>
+    );
 }
 
 export default SetupViewContainer;
